writer: Split the 'writer.Writer' implementation in different files
This way it's easy to read and know what's what
